requirements.txt
python-telegram-bot>=21.0
openai>=1.0
pymupdf4llm>=0.0.10
pymupdf>=1.24.0
markdown>=3.5
weasyprint>=60.0
python-dotenv>=1.0
Pillow>=10.0
ebooklib>=0.18
html2text>=2024.2.26
